Change clang-format alignment rules
Rationale: To preserve the functionality of tools like `git blame`, try
not to touch any lines that are not affected by the change due to
formatting rules.

This means that horizontal alignment is disabled, because adding a line
that needs wider alignment to a block requires re-aligning the entire
block.